requirements.txt
Flask
tensorflow-cpu
opencv-python-headless
Pillow
numpy
gunicorn
